backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
pdfplumber
opencv-python
numpy
python-docx
PyPDF2

